ollama-agent/requirements.txt
google-adk
pdfplumber
litellm